flask_jwt_app/requirements.txt
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
distlib==0.3.9
filelock==3.18.0
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pipenv==2024.4.1
platformdirs==4.3.7
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.0
setuptools==79.0.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
virtualenv==20.30.0
Werkzeug==3.1.3
